refactor(global-search): move useDebounceValue to utils
