+++ plugins/global-search/src/utils/useDebounceValue.tsx
@@ -1,0 +1,16 @@
+import { useEffect, useState } from "react"
+
+export function useDebounceValue<T>(value: T, delay: number) {
+    const [debouncedValue, setDebouncedValue] = useState<T>(value)
+
+    useEffect(() => {
+        const timeout = setTimeout(() => {
+            setDebouncedValue(value)
+        }, delay)
+        return () => {
+            clearTimeout(timeout)
+        }
+    }, [value, delay])
+
+    return debouncedValue
+}
